gh-4019 Move ensureFullStop to stroom.util.string.StringUtil
Cos GWT
