+++ stroom-util/src/main/java/stroom/util/string/StringUtil.java
@@ -89,4 +89,19 @@ public class StringUtil {
         }
         return stringBuilder.toString();
     }
+
+    public static String ensureFullStop(final String str) {
+        if (str == null) {
+            return "";
+        } else {
+            final String str2 = str.stripTrailing();
+            if (str2.isBlank()) {
+                return str;
+            } else if (str2.endsWith(".")) {
+                return str2;
+            } else {
+                return str2 + ".";
+            }
+        }
+    }
 }
