+++ stroom-util/src/test/java/stroom/util/string/TestStringUtil.java
@@ -207,4 +207,20 @@ class TestStringUtil {
                 .addCase(500, expectedOutput)
                 .build();
     }
+
+    @TestFactory
+    Stream<DynamicTest> testEnsureFullStop() {
+        return TestUtil.buildDynamicTestStream()
+                .withInputAndOutputType(String.class)
+                .withSingleArgTestFunction(StringUtil::ensureFullStop)
+                .withSimpleEqualityAssertion()
+                .addCase(null, "")
+                .addCase("", "")
+                .addCase(" ", " ")
+                .addCase("foo", "foo.")
+                .addCase("foo ", "foo.")
+                .addCase("foo.", "foo.")
+                .addCase("foo. ", "foo.")
+                .build();
+    }
 }
